ikea: Move to own pkg
